requirements.txt
Pillow==3.1.0
requests==2.9.1
Shapely==1.5.13
tiletanic==0.0.4
wheel==0.24.0
